Add dedent() function, which dedents and removes newlines from strings so that you can more easily write multiline error message etc, such as: ``` def foo()     msg = """There was an error, which we have written in code         over multiple lines, but with indentation. This indentation         and these line breaks should be ignored.
            1. But this extra indentation
            2. and line endings
            3. should be included verbatim

        and then this text at the original indentation level should be
        unwrapped again.
        """
    raise ValueError(dedent(msg))
```

This function is not quite the same as Python's textwrap.dedent, as this one
excludes the first line from the indentation calculation, removes newlines,
and includes lines with additional indentation verbatim. I've been using it to
make long error messages without having to manually join strings with '+' for
each line, which plays well with auto-wrapping text in text editors, e.g.
alt-Q in sublime text.