Allow serialiser to serialise bytestings, by base64 encoding them with a prefix.
JSON does not distinguish between bytestrings and unicode strings, but Python
does, and Python 3 has stricter requirements not to confuse the two.

In order for a bytestrings to survive serialisation and deserialisation, this
change encodes bytestrings as base64, adds a prefix, and the result is serialised
in the JSON ouput the same as any unicode string. Upon deserialisation, strings
are inspected for the prefix and decoded.

This is motivated by the labscript driver for NI IMAQdx cameras. Camera attributes
can be set, which are currently stored as connection table properties or device
properties. Currently they are stored as strings, even though this leads to
various issues with the Python nivision API choking on some of them since they
are not the right datatype. Not only can attributes be strings or integers or floats
etc, but nivision makes a distinction between byestrings and unicode strings.
So to correctly set these, our serialisation needs to allow this distinction to
round-trip.

This encoding of bytestrings is only performed for Python 3. The NI IMAQdx camera
server is Python3-only in any case.